% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_time.R
\name{efficiency_graphics_mem}
\alias{efficiency_graphics_mem}
\title{HPC efficiency graphics}
\usage{
efficiency_graphics_mem(eff_table, interactive = TRUE)
}
\arguments{
\item{eff_table}{data.frame from efficiency_table function}

\item{interactive}{logical indicating if a plotly graphic should be displayed}
}
\description{
HPC efficiency graphics
}
