% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_time.R
\name{efficiency_graphics_mem_cpu}
\alias{efficiency_graphics_mem_cpu}
\title{HPC efficiency graphics}
\usage{
efficiency_graphics_mem_cpu(eff_table, reescale = 15)
}
\description{
HPC efficiency graphics
}
