% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_time.R
\name{efficiency_table}
\alias{efficiency_table}
\title{Convert data obtained by efficiency script to
tables}
\usage{
efficiency_table(infos_seff, jobs)
}
\arguments{
\item{infos_seff}{information obtained throught command seff}

\item{jobs}{vector defining job numbers}
}
\description{
Convert data obtained by efficiency script to
tables
}
