% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_simu.R
\name{errorProb_graph}
\alias{errorProb_graph}
\title{Functions to build graphics for the data from simulations}
\usage{
errorProb_graph(data, genotypes, alpha)
}
\description{
Functions to build graphics for the data from simulations
}
