% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_emp.R
\name{errorProb_graph_emp}
\alias{errorProb_graph_emp}
\title{Functions to build graphics for the empirical datas}
\usage{
errorProb_graph_emp(data, genotypes, from)
}
\description{
Functions to build graphics for the empirical datas
}
