% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_emp.R
\name{prepare_datas_emp}
\alias{prepare_datas_emp}
\title{Read empirical input data}
\usage{
prepare_datas_emp(x, example_emp)
}
\arguments{
\item{x}{inputted dataset}

\item{example_emp}{selected example id}
}
\description{
Read empirical input data
}
