% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_simu.R
\name{prepare_datas_simu}
\alias{prepare_datas_simu}
\title{Read input data}
\usage{
prepare_datas_simu(x, example_simu)
}
\arguments{
\item{x}{inputted dataset}

\item{example_simu}{selected example id}
}
\description{
Read input data
}
