% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_poly_emp.R
\name{prepare_poly_datas_emp}
\alias{prepare_poly_datas_emp}
\title{Read empirical input polyploid data}
\usage{
prepare_poly_datas_emp(x = NULL, example_emp = NULL)
}
\arguments{
\item{x}{inputted dataset}

\item{example_emp}{selected example id}
}
\description{
Read empirical input polyploid data
}
