% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_emp.R
\name{SNPCalling_efficiency_graph_emp}
\alias{SNPCalling_efficiency_graph_emp}
\title{Performs comparisons between maximum datasets:
\enumerate{
\item Venn diagram with exactly matching markers positions;
}}
\usage{
SNPCalling_efficiency_graph_emp(data, data_names)
}
\arguments{
\item{data_names}{character defining the name of each vcf data}

\item{data_list}{list with marker names (chr_pos) of each dataset}

\item{out_prefix}{character defining the prefix of image files names}
}
\description{
Performs comparisons between maximum datasets:
\enumerate{
\item Venn diagram with exactly matching markers positions;
}
}
