% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_time.R
\name{efficiency_table}
\alias{efficiency_table}
\title{Convert data obtained by efficiency script to
tables}
\usage{
efficiency_table(infos_seff, jobs)
}
\arguments{
\item{infos_seff}{output of seff bash command in HPRC}

\item{jobs}{data.frame with first column being the task name and the second the job id}
}
\description{
Convert data obtained by efficiency script to
tables
}
