% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointSeeds.R
\name{joint_same_depth}
\alias{joint_same_depth}
\title{Joint simulations with same depth and different seed}
\usage{
joint_same_depth(..., out_name = "SimulatedReads_results_joint")
}
\arguments{
\item{...}{tar.gz files outputted from onemap_workflows}

\item{out_name}{output file name without the extension definition}
}
\description{
Joint simulations with same depth and different seed
}
